using classes for services.
Take care new isntance is saved in memory while the script or server is
running.
This gives an memory persistence idea.
This because in products.router the class is geneate as a global
variable, while all the mehtod are rerunning by the app of express..;
The service class was created an stored like a global variable.